Refactor: Create class Registerable; Make file for class Query.
